pyproject.toml
[project]
name = "api-hcc-plan"
version = "0.1.0"
description = "FastAPI-basierte Webanwendung für HCC-Planung mit HTMX und OAuth2"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pony>=0.7.17",
    "jinja2>=3.1.3",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "ruff>=0.1.11",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]